+++ src/main/java/io/github/riverson98/customerBase/repository/PersonRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PersonRepository extends JpaRepository<ClientEntity, Long> {
+public interface PersonRepository extends JpaRepository<ClientEntity, Integer> {
 }
